Update 6k.html
I have added some extra line of code where I put conditional operators so that the cart can't be on negative figure. Which is illogical. I have learnt the basic of Javascript and the logical aspects of coding I hope that will help all to understand the lag of this code. Correct me If I wrong